requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.1
PyYAML==6.0.1
colorlog==6.9.0
rich==13.9.4
Jinja2==3.1.3
aiohttp>=3.8.0